+++ src/components/home/Features/PaidSection/PaidSection.tsx
